+++ .editorconfig
@@ -1,0 +1,4 @@
+[*]
+charset = utf-8
+indent_size = 2
+indent_style = space
