+++ .gitignore
@@ -1,3 +1,4 @@
 /.idea/
 /vendor/
 /composer.lock
+/build.png
