+++ src/ClientException.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace Jawira\PlantUmlClient;
+
+
+use Exception;
+
+class ClientException extends Exception
+{
+
+}
