--- src/PlantUmlClientException.php
@@ -1,12 +1,0 @@
-<?php
-
-
-namespace Jawira\PlantUmlRequest;
-
-
-use Exception;
-
-class PlantUmlClientException extends Exception
-{
-
-}
